projects/todo-app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
PyJWT>=2.8.0

# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
